Messy first step to deferred execution
The idea is that the use should be able to pick an invocation strategy that
yields functions that are called in two phases: first parameter extraction,
then execution. By triggering the second step themselves, the user is able to
e.g. do parameter extraction for multiple functions and only proceed to
execution if all extractions are successful.

Here for function return type R we return () => R.
But we want to end up in a position to supporting things like
Either[List[String], () => R], with a Left representing parameter extraction
errors.